Fix DZR_CBC extraction on *BSD like
The grep command for the DZR_CBC extract use gnu extension.

P and o arguments are not POSIX and have a different meaning on BSD system.
You have to install gnugrep and use it directly.
Either by the g prefixed version on OpenBSD, NetBSD and macOs(Darwin) or
the standard PATH for port on FreeBSD.